Fixed build for mesa and repo initialization

+++ build-rpms.sh
@@ -21,6 +21,7 @@ for dir in ./*/; do
   ${SCRIPT} "${1}" "${BUILDDIR}" "${RESULTDIR}" || true
 done
 [[ -d "${RESULTDIR}/${ARCH}" ]] || mkdir -p ${RESULTDIR}/{aarch64,x86_64,ppc64le}/{debug/tree,os} ${RESULTDIR}/source/tree
+# TODO: sign rpms
 mv ${BUILDDIR}/*.src.rpm "${RESULTDIR}/source/tree/"
 mv ${BUILDDIR}/*-debug{info,source}-*.rpm "${RESULTDIR}/${ARCH}/debug/tree/"
 cp ${BUILDDIR}/*.noarch.rpm "${RESULTDIR}/aarch64/os"
@@ -28,7 +29,7 @@ cp ${BUILDDIR}/*.noarch.rpm "${RESULTDIR}/x86_64/os"
 cp ${BUILDDIR}/*.noarch.rpm "${RESULTDIR}/ppc64le/os"
 rm ${BUILDDIR}/*.noarch.rpm
 mv ${BUILDDIR}/*.rpm "${RESULTDIR}/${ARCH}/os"
-for repo in "${RESULTDIR}/source/tree/" "${RESULTDIR}/${ARCH}/debug/tree/" "${RESULTDIR}/${ARCH}/os"; do
-  rm -f $(repomanage --old ${repo})
+for repo in "${RESULTDIR}/source/tree" "${RESULTDIR}/${ARCH}/debug/tree" "${RESULTDIR}/${ARCH}/os"; do
+  [[ -d "${repo}/repodata" ]] && rm -f $(repomanage --old ${repo})
   createrepo ${repo}
 done
